backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.6.0
flask-cors==4.0.1
sqlalchemy==2.0.35
cryptography==42.0.8
openai==1.51.0
python-dotenv==1.0.1
bcrypt==4.2.0
marshmallow==3.22.0
requests==2.32.3
apscheduler==3.10.4
pytest==8.3.3
pytest-flask==1.3.0
